Move to people app, expose "add new contact"
