+++ frontend/src/people/People.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { PersonForm } from "./PeopleForms";
+import { OwnerContactForm, PersonForm } from "./PeopleForms";
 import { PersonView } from "./PeopleViews";
 
 const header_style = {
@@ -34,5 +34,17 @@ export const UpdateReporter = ({id}) => (
 export const ReporterDetail = ({id}) => (
   <div>
     <PersonView id={id} />
+  </div>
+)
+
+export const NewOwnerContact = () => (
+  <div>
+    <OwnerContactForm />
+  </div>
+)
+
+export const UpdateOwnerContact = ({id}) => (
+  <div>
+    <OwnerContactForm id={id} />
   </div>
 )
