+++ hotline/urls.py
@@ -7,7 +7,6 @@ app_name = 'hotline'
 router = DefaultRouter()
 router.register(r'servicerequests', views.ServiceRequestViewSet)
 router.register(r'visitnote', views.VisitNoteViewSet)
-router.register(r'ownercontact', views.OwnerContactViewSet)
 
 urlpatterns = [
     path('api/', include(router.urls)),
