+++ people/views.py
@@ -3,8 +3,8 @@ from actstream import action
 
 from animals.models import Animal
 from hotline.models import ServiceRequest
-from people.models import Person
-from people.serializers import PersonSerializer
+from people.models import OwnerContact, Person
+from people.serializers import OwnerContactSerializer, PersonSerializer
 
 
 # Provides view for Person API calls.
@@ -43,3 +43,9 @@ class PersonViewSet(viewsets.ModelViewSet):
         if serializer.is_valid():
             person = serializer.save()
             action.send(self.request.user, verb='updated person', target=person)
+
+class OwnerContactViewSet(viewsets.ModelViewSet):
+
+    queryset = OwnerContact.objects.all()
+    permission_classes = [permissions.IsAuthenticated, ]
+    serializer_class = OwnerContactSerializer
